latest version

+++ src/diskarray.h
@@ -2,6 +2,7 @@
 #define _DISKARRAY_H
 #include <cstdio>
 #include <string>
+#include "constants.h"
 using namespace std;
 class diskArray
 {
@@ -10,24 +11,24 @@ class diskArray
 	FILE * fp;
 
 	//the row and column number
-        int rows;
+        long long rows;
         int length;
 
 	//in memory cache
 	int * cached_row;
 
 	//the current row number;
-	int cur_row;
+	long long cur_row;
 
 	//flush the cache to disk;
 	void flush();
 
 	//load a row from the disk
-	void load(int rows);
+	void load(long long rows);
 public:
 	//set the row and column number
         void init(string name);
-	void set_length(int rownumber, int len);
+	void set_length(long long rownumber, int len);
 
 	//visit an element
 	int &visit(int i,int j);
